+++ stalky106.sln
@@ -3,15 +3,18 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 16
 VisualStudioVersion = 16.0.28803.156
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "stalky106", "Stalky106\stalky106.csproj", "{B9FBF509-2FA5-4A68-907C-A1A016225EC9}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Stalky106", "Stalky106\Stalky106.csproj", "{B9FBF509-2FA5-4A68-907C-A1A016225EC9}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Release|x64 = Release|x64
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{B9FBF509-2FA5-4A68-907C-A1A016225EC9}.Release|x64.ActiveCfg = Release|x64
-		{B9FBF509-2FA5-4A68-907C-A1A016225EC9}.Release|x64.Build.0 = Release|x64
+		{B9FBF509-2FA5-4A68-907C-A1A016225EC9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{B9FBF509-2FA5-4A68-907C-A1A016225EC9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{B9FBF509-2FA5-4A68-907C-A1A016225EC9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{B9FBF509-2FA5-4A68-907C-A1A016225EC9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
